requirements.txt
torch
torchvision
diffusers>=0.27.0
transformers
accelerate
peft
Pillow
huggingface_hub  # for hf_hub_download used by the Anima pipeline
fastapi
uvicorn[standard]
